EXACT: ProfileDeltaFrame, IsInHardware, SetDimensions, EnterMode, ~CPathHazard
ctor, ProcessPal, IsPlaying, IsLooping - plus ProcessPalBGR (95.6),
ProcessPalQuad (62.7), CGruntzSoundZ::Init, CRezDirNode::Load and
DirectSoundMgr::Play falling out of the same discoveries.

A REAL BUG that `--diff` was hiding: CDDrawSurfaceMgr::SetDimensions read as
"identical asm" because the diff masks BRANCH DISPLACEMENTS. Retail's `74 2e`
skips the m_level rebuild; ours was `74 16` and fell into it - we rebuilt the
whole level on every no-op resize. The rebuild belongs inside the resize guard.

Two more: CGruntzSoundZ::Init's third arg SUPPRESSES midi rather than skipping
bring-up (retail's `jne` lands on the shared m_enabled = 0 store, not on
`return 1`) - renamed noMidi. And four unsigned counters recovered from ja/jbe
guards (RezSrc::m_1c, four CSymParser fields in both the in-memory and packed
on-disk structs, CAniSource::m_namelen, CMapMgr::AllocGrid's row counter),
which also kills 8 casts and the last (i32)strlen.

Identity: CPathHazard's leg/strike deadline+window pairs are a real sub-object.
Retail batches the i64 zero-stores lo,lo,hi,hi per two i64s, which ONLY a nested
struct with its own default ctor produces - chained assignment, mem-init list,
inline helpers, block scopes and a shared i64 temp were all tried and fail.
New struct CHazardTimer; corroborated by the serializer already streaming each
pair through one helper.

THE CARRY-FORWARD - the jcc-sequence sieve. Diff the ordered
(offset, mnemonic, target) branch list of base vs target per function; equal
count with 1-4 differing MNEMONICS at matching positions is the signal, and
displacement shifts are noise. Swept all 1309 sub-100% functions: 105 hits,
6 SIGNEDNESS and 53 pure je/jne polarity, both halves proven actionable (4
signedness fixes and 1 polarity fix landed, 2 straight to EXACT).
docs/patterns/masked-diff-hides-branch-target.md.

Two parked with the sweep recorded in-source rather than a hand-wave:
BlitShadeNorm (nine spellings, all byte-identical prologue - cl reassociates the
sub chain before regalloc) and RemoveValue (13 spellings; every key-first form
does move the CString into the reclaimed slot but then emits the ctor before the
GetStartPosition neg/sbb - strictly worse).

exact 2972 (69.7%), casts 440, 77.72% fuzzy. `gruntz build --full` exit 0.

+++ src/Bute/SymTab.cpp
@@ -9,8 +9,8 @@
 // (Load@CRezDirNode 0x13a0f0; @identity-TODO). 0x13c080 is Classify@CSymParser
 // (its callers ParseBuffer/LoadEntry pass ecx=this; ex the FindEntry@CRezDir guess).
 #include <Rez/DebugPrintf.h> // RezAssertFail (owning decl; varargs C linkage)
-#include <Bute/SymTab.h> // own extern surface
-#include <Mfc.h>         // afx-first (RezMgr.h below pulls MFC/Win32 for the two Rez strays)
+#include <Bute/SymTab.h>     // own extern surface
+#include <Mfc.h>             // afx-first (RezMgr.h below pulls MFC/Win32 for the two Rez strays)
 #include <rva.h>
 #include <io.h>     // _finddata_t / _findfirst / _findnext / _findclose (ParseRecords)
 #include <stdlib.h> // _splitpath (0x18c530) / atoi (0x11ff10)
@@ -204,8 +204,7 @@ char* CParseSource::BeginParse() {
     if (m_buffer == 0) {
         return 0;
     }
-    if (m_reader->Read(m_base, 0, m_length, m_buffer)
-        != static_cast<i32>(m_length)) {
+    if (m_reader->Read(m_base, 0, m_length, m_buffer) != static_cast<i32>(m_length)) {
         ::operator delete(m_buffer);
         m_buffer = 0;
     }
@@ -597,7 +596,7 @@ CSymTab* CSymTab::CreateSub(const char* name) {
         return 0;
     }
     m_subTabs.Insert(&child->m_node20);
-    if (m_owner->m_longestScopeNameLen <= static_cast<i32>(strlen(name))) {
+    if (m_owner->m_longestScopeNameLen <= strlen(name)) {
         m_owner->m_longestScopeNameLen = strlen(name) + 1;
     }
     return child;
@@ -641,15 +640,14 @@ CParseSource* CSymTab::AddNamedValue(void* a1, void* name, i32 key) {
     }
     rec->m_valTable.Insert(&slot->m_node1c);
     u32 len = strlen(static_cast<char*>(name));
-    if (static_cast<u32>(m_owner->m_longestLeafNameLen) <= len) {
+    if (m_owner->m_longestLeafNameLen <= len) {
         m_owner->m_longestLeafNameLen = len + 1;
     }
     return slot;
 }
 
 RVA(0x0013a4b0, 0x75)
-CParseSource*
-CSymTab::AddNodeEntry(u32 key, const char* name, CSymRec* rec, CRezItmBase* stream) {
+CParseSource* CSymTab::AddNodeEntry(u32 key, const char* name, CSymRec* rec, CRezItmBase* stream) {
     CParseSource* slot = m_owner->PopParseSlot();
     if (slot == 0) {
         return slot;
@@ -672,7 +670,7 @@ CSymTab::AddNodeEntry(u32 key, const char* name, CSymRec* rec, CRezItmBase* stre
     );
     rec->m_valTable.Insert(&slot->m_node1c);
     u32 len = strlen(name);
-    if (static_cast<u32>(m_owner->m_longestLeafNameLen) <= len) {
+    if (m_owner->m_longestLeafNameLen <= len) {
         m_owner->m_longestLeafNameLen = len + 1;
     }
     return slot;
@@ -697,7 +695,7 @@ i32 CSymTab::AddNodeSubEntry(void* rec, void* found) {
 // constant in ebp; the recompile swaps them (a2->ebp, 0->ebx), which cascades through
 // every null check + the recursion fourcc setup. Banked for the final sweep.
 RVA(0x0013a580, 0xb2)
-i32 CSymTab::ApplyRecursive(CRezItmBase * a0, i32 a1, i32 a2, i32 a3) {
+i32 CSymTab::ApplyRecursive(CRezItmBase* a0, i32 a1, i32 a2, i32 a3) {
     i32 ok = 1;
     if (a2 != 0) {
         CHashElement* e = m_subTabs.First();
@@ -974,8 +972,7 @@ CSymParser::~CSymParser() {
         Clear(0);
     }
     CRezItmBase* p;
-    for (p = m_list.m_head; p != 0;
-         p = m_list.m_head) {
+    for (p = m_list.m_head; p != 0; p = m_list.m_head) {
         m_list.Remove(p);
         m_list.m_count--;
         delete p; // the slot-1 scalar-deleting dtor (delete emits the same null test)
@@ -1203,20 +1200,20 @@ i32 CSymParser::LoadEntry(char* name, i32 flag) {
     SymTabFileHeader hdr;
     node->Read(0, 0, 0xa8, &hdr);
     u32 v;
-    v = static_cast<u32>(hdr.m_54);
-    if (v > static_cast<u32>(m_54)) {
+    v = hdr.m_54;
+    if (v > m_54) {
         m_54 = v;
     }
-    v = static_cast<u32>(hdr.m_longestScopeNameLen);
-    if (v > static_cast<u32>(m_longestScopeNameLen)) {
+    v = hdr.m_longestScopeNameLen;
+    if (v > m_longestScopeNameLen) {
         m_longestScopeNameLen = v;
     }
-    v = static_cast<u32>(hdr.m_longestLeafNameLen);
-    if (v > static_cast<u32>(m_longestLeafNameLen)) {
+    v = hdr.m_longestLeafNameLen;
+    if (v > m_longestLeafNameLen) {
         m_longestLeafNameLen = v;
     }
-    v = static_cast<u32>(hdr.m_60);
-    if (v > static_cast<u32>(m_60)) {
+    v = hdr.m_60;
+    if (v > m_60) {
         m_60 = v;
     }
     m_root->ApplyRecursive(node, hdr.m_scopeCount, hdr.m_leafCount, flag);
@@ -1340,8 +1337,7 @@ i32 CSymParser::Clear(i32 final) {
     delete m_activeNode; // slot-1 scalar dtor (delete emits the same null test)
     m_activeNode = 0;
     CRezItmBase* p;
-    for (p = m_list.m_head; p != 0;
-         p = m_list.m_head) {
+    for (p = m_list.m_head; p != 0; p = m_list.m_head) {
         p->Close();
         m_list.Remove(p);
         m_list.m_count--;
